Change HashBase.dumpData return type to always include a partition key
Previously, files that are not included in a partition would be written
to the data dump at the same level as partitioned datasets. This change
is a fix in a way; it makes the data structure more sensible with which
to work.

BREAKING_CHANGE: Return type for HashMap.dataDump has a slight semantic
change.