Finish memlet integration and upgrade to v0.1.x
Finish the half baked work where some parts of the implementation use a
disklet instance and other parts use a memlet instance.
All baselets use memlets internally whether given a memlet or not.

An upgrade to the latest version of memlet is included in order to
access new APIs necessary to finish integration (i.e. `navigateMemlet`).